move magicMetadata types to its own type file
